requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain-huggingface
langchain-core
langgraph
pydantic
huggingface_hub
tzdata